pyproject.toml
[project]
name = "finradar"
version = "0.1.0"
description = "AI-powered financial intelligence agent with scheduled report acquisition and startup discovery"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "FinRadar Team", email = "team@finradar.ai"}
]
keywords = ["ai", "agent", "finance", "sec", "startups", "dspy", "pydantic", "prefect"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.1.0",
    "pydantic-ai>=0.0.1",

    "dspy-ai>=2.5.0",
    "llama-index>=0.12.0",
    "llama-index-core>=0.12.0",
    "llama-index-readers-file>=0.4.0",
    "llama-index-readers-web>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-llms-anthropic>=0.4.0",

    "prefect>=3.0.0",
    "prefect-aws>=0.5.0",

    "anthropic>=0.40.0",
    "openai>=1.60.0",

    "httpx>=0.27.0",
    "aiofiles>=23.2.0",
    "tenacity>=9.0.0",
    "structlog>=25.0.0",

    "sec-edgar-downloader>=5.0.0",
    "yfinance>=0.2.50",
    "alpha-vantage>=2.3.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",

    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pdfplumber>=0.11.0",

    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    "redis>=5.2.0",

    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",

    "typer>=0.15.0",
    "rich>=13.9.0",

    "qdrant-client>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "black>=25.0.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
finradar = "app.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]
